Adding more tests
